requirements.txt
Flask==2.2.2
pyserial==3.5
